fix: conditional compiling for cgroups code
