+++ pkg/drivers/kic/oci/cgroups_linux.go
@@ -19,9 +19,15 @@ limitations under the License.
 package oci
 
 import (
+	"fmt"
+	"os"
+	"path"
 	"syscall"
 
+	"github.com/opencontainers/runc/libcontainer/cgroups"
 	"golang.org/x/sys/unix"
+
+	"k8s.io/klog/v2"
 )
 
 // IsCgroup2UnifiedMode returns whether we are running in cgroup 2 cgroup2 mode.
@@ -31,4 +37,57 @@ func IsCgroup2UnifiedMode() (bool, error) {
 		return false, err
 	}
 	return st.Type == unix.CGROUP2_SUPER_MAGIC, nil
+}
+
+// findCgroupMountpoints returns the cgroups mount point
+// defined in docker engine engine/pkg/sysinfo/sysinfo_linux.go
+func findCgroupMountpoints() (map[string]string, error) {
+	cgMounts, err := cgroups.GetCgroupMounts(false)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to parse cgroup information: %v", err)
+	}
+	mps := make(map[string]string)
+	for _, m := range cgMounts {
+		for _, ss := range m.Subsystems {
+			mps[ss] = m.Mountpoint
+		}
+	}
+	return mps, nil
+}
+
+// HasMemoryCgroup checks whether it is possible to set memory limit for cgroup.
+func HasMemoryCgroup() bool {
+	cgMounts, err := findCgroupMountpoints()
+	if err != nil {
+		klog.Warning("Your kernel does not support memory limit capabilities or the cgroup is not mounted.")
+		return false
+	}
+	_, ok := cgMounts["memory"]
+	if !ok {
+		klog.Warning("Your kernel does not support memory limit capabilities or the cgroup is not mounted.")
+		return false
+	}
+	return true
+}
+
+// hasMemorySwapCgroup checks whether it is possible to set swap limit for cgroup
+func hasMemorySwapCgroup() bool {
+	cgMounts, err := findCgroupMountpoints()
+	if err != nil {
+		klog.Warning("Your kernel does not support swap limit capabilities or the cgroup is not mounted.")
+		return false
+	}
+	mountPoint, ok := cgMounts["memory"]
+	if !ok {
+		klog.Warning("Your kernel does not support swap limit capabilities or the cgroup is not mounted.")
+		return false
+	}
+
+	_, err = os.Stat(path.Join(mountPoint, "memory.memsw.limit_in_bytesw"))
+	if err != nil {
+		klog.Warning("Your kernel does not support swap limit capabilities or the cgroup is not mounted.")
+		return false
+
+	}
+	return true
 }
